ibi mgmt: add ibbf test case
