+++ tests/lca/imagebasedinstall/mgmt/deploy/internal/tsparams/const.go
@@ -6,4 +6,7 @@ const (
 
 	// LabelEndToEndDeployment represents e2e label that can be used for test cases selection.
 	LabelEndToEndDeployment = "e2e"
+
+	// LabelReinstall represents reinstall label that can be used for test cases selection.
+	LabelReinstall = "reinstall"
 )
