+++ tests/lca/imagebasedinstall/mgmt/internal/mgmtconfig/config.go
@@ -58,20 +58,23 @@ type Cluster struct {
 }
 
 // MGMTConfig type contains mgmt configuration.
+//
+//nolint:lll
 type MGMTConfig struct {
 	*ibiconfig.IBIConfig
-	Cluster          *Cluster
-	ClusterInfoPath  string `envconfig:"ECO_LCA_IBI_MGMT_CLUSTER_INFO"`
-	SeedImage        string `envconfig:"ECO_LCA_IBI_MGMT_SEED_IMAGE" default:"quay.io/ocp-edge-qe/ib-seedimage-public:ci"`
-	SeedClusterInfo  *seedimage.SeedImageContent
-	SSHKeyPath       string `envconfig:"ECO_LCA_IBI_MGMT_SSHKEY_PATH"`
-	PublicSSHKey     string
-	StaticNetworking bool   `envconfig:"ECO_LCA_IBI_MGMT_STATIC_NETWORK" default:"false"`
-	ExtraManifests   bool   `envconfig:"ECO_LCA_IBI_EXTRA_MANIFESTS" default:"true"`
-	CABundle         bool   `envconfig:"ECO_LCA_IBI_CA_BUNDLE" default:"true"`
-	SiteConfig       bool   `envconfig:"ECO_LCA_IBI_SITECONFIG" default:"true"`
-	ExtraPartName    string `envconfig:"ECO_LCA_IBI_MGMT_EXTRA_PARTITION_NAME" default:""`
-	ExtraPartSizeMib string `envconfig:"ECO_LCA_IBI_MGMT_EXTRA_PARTITION_SIZE" default:"50000"`
+	Cluster              *Cluster
+	ClusterInfoPath      string `envconfig:"ECO_LCA_IBI_MGMT_CLUSTER_INFO"`
+	SeedImage            string `envconfig:"ECO_LCA_IBI_MGMT_SEED_IMAGE" default:"quay.io/ocp-edge-qe/ib-seedimage-public:ci"`
+	SeedClusterInfo      *seedimage.SeedImageContent
+	SSHKeyPath           string `envconfig:"ECO_LCA_IBI_MGMT_SSHKEY_PATH"`
+	PublicSSHKey         string
+	StaticNetworking     bool   `envconfig:"ECO_LCA_IBI_MGMT_STATIC_NETWORK" default:"false"`
+	ExtraManifests       bool   `envconfig:"ECO_LCA_IBI_EXTRA_MANIFESTS" default:"true"`
+	CABundle             bool   `envconfig:"ECO_LCA_IBI_CA_BUNDLE" default:"true"`
+	SiteConfig           bool   `envconfig:"ECO_LCA_IBI_SITECONFIG" default:"true"`
+	ExtraPartName        string `envconfig:"ECO_LCA_IBI_MGMT_EXTRA_PARTITION_NAME" default:""`
+	ExtraPartSizeMib     string `envconfig:"ECO_LCA_IBI_MGMT_EXTRA_PARTITION_SIZE" default:"50000"`
+	ReinstallManifestDir string `envconfig:"ECO_LCA_IBI_REINSTALL_MANIFEST_DIR" default:""`
 }
 
 // NewMGMTConfig returns instance of MGMTConfig type.
